+++ src/GuardClauses/GuardValidationObjectExtensions.cs
@@ -1,0 +1,27 @@
+﻿using System.Runtime.CompilerServices;
+using PowerUtils.Validations.Exceptions;
+
+namespace PowerUtils.Validations.GuardClauses
+{
+    public static class GuardValidationObjectExtensions
+    {
+        /// <summary>
+        /// Throws an <see cref="PropertyException" /> if <paramref name="value"/> is null with the error code 'REQUIRED'
+        /// </summary>
+        /// <param name="_"></param>
+        /// <param name="value">Value to validate</param>
+        /// <param name="parameterName">If not defined, the name of the variable passed by the <paramref name="value"/> parameter will be used</param>
+        /// <exception cref="PropertyException">Exception thrown when value is null</exception>
+        public static void IfNull(
+            this IGuardValidationClause _,
+            object value,
+            [CallerArgumentExpression("value")] string parameterName = null
+        )
+        {
+            if(value == null)
+            {
+                throw new PropertyException(parameterName, ErrorCodes.REQUIRED);
+            }
+        }
+    }
+}